+++ src/smurf_f.h
@@ -159,7 +159,7 @@ void    f_init_bintable(OBJECT *rsc);       /* Bin„rwert-Table init */
 
 
 int handle_keyboardevent(WINDOW *wind_s, int scancode, int *sel_object);
-void change_object(WINDOW *window, int objct, int status, int redraw);
+void change_object(WINDOW *window, WORD objct, WORD status, int redraw);
 
 
 /*
